+++ src/commands/changelog/index.ts
@@ -3,8 +3,13 @@ import api from "api"
 import { SlashCommand } from "types/common"
 import { composeEmbedMessage } from "ui/discord/embed"
 import UI, { Platform } from "@consolelabs/mochi-formatter"
-import mochiAPI from "../../adapters/mochi-api"
-import profile from "../../adapters/profile"
+import mochiAPI from "adapters/mochi-api"
+import profile from "adapters/profile"
+import { HOMEPAGE_URL } from "utils/constants"
+import { ButtonInteraction, Message } from "discord.js"
+import { getAuthor, getEmojiURL, emojis } from "utils/common"
+import { WHITE_LIST_PUBLIC_CHANGELOG } from "env"
+import { APIError } from "../../errors"
 
 const slashCmd: SlashCommand = {
   name: "changelog",
@@ -16,7 +21,15 @@ const slashCmd: SlashCommand = {
     return data
   },
   run: async function (i) {
-    const changelog = api.getLatestChangelog()
+    const { data } = await api.base.metadata.getChangelogs()
+    const changelogs =
+      data?.sort((a, b) => {
+        const timeA = new Date(a.created_at).getTime()
+        const timeB = new Date(b.created_at).getTime()
+        return timeB - timeA
+      }) ?? []
+
+    const changelog = changelogs.find((c) => !c.is_expired) ?? null
     if (!changelog) {
       return {
         messageOptions: {
@@ -38,19 +51,21 @@ const slashCmd: SlashCommand = {
       return
     }
 
-    const { text } = await UI.components.changelog({
+    const { text, images } = await UI.components.changelog({
       title: changelog.title,
       content: changelog.content,
       on: Platform.Discord,
     })
-    const sections = text.split("<br>")
-    for (let j = 0; j < sections.length; j++) {
-      if (j === 0) {
-        await i.editReply(sections[j])
-        continue
-      }
-      await i.channel?.send(`${sections[j]}`)
+    const footer = `\⎯⎯⎯⎯⎯\nView changelog detail on [Mochi web](${HOMEPAGE_URL}/changelog/${changelog.version})`
+    let embed = composeEmbedMessage(null, {
+      description: `${text}\n${footer}`,
+      footer: [`v${changelog.version}`],
+    })
+    if (images.length > 0) {
+      embed.setImage(images[0])
     }
+
+    await i.editReply({ embeds: [embed] })
 
     return null
   },
@@ -61,3 +76,41 @@ const slashCmd: SlashCommand = {
 }
 
 export default { slashCmd }
+
+export async function handleConfirmPublicChangelog(i: ButtonInteraction) {
+  const args = i.customId.split("_")
+  const changelogName = args[3]
+
+  const author = getAuthor(i)
+  if (!WHITE_LIST_PUBLIC_CHANGELOG.includes(author.id)) {
+    return
+  }
+
+  await i.deferReply({ ephemeral: true })
+
+  const data = await mochiAPI.publicChangelog(changelogName, true)
+  if (!data.ok) {
+    throw new APIError({
+      msgOrInteraction: i,
+      description: "Cannot publish changelog",
+      curl: data.curl,
+      error: data.error ?? "",
+      status: data.status ?? 500,
+    })
+  }
+
+  await i.editReply({
+    embeds: [
+      composeEmbedMessage(null, {
+        author: ["Publish Changelog successfully", getEmojiURL(emojis.APPROVE)],
+        description: `You published the new changelog`,
+      }),
+    ],
+  })
+
+  const { message } = <{ message: Message }>i
+
+  await message.edit({ embeds: i.message.embeds, components: [] })
+
+  return
+}