+++ src/types/api.ts
@@ -424,11 +424,16 @@ export interface ModelProductChangelogs {
   created_at?: string
   file_name?: string
   github_url?: string
+  id?: number
   is_expired?: boolean
+  next_version?: string
+  previous_version?: string
   product?: string
+  seo_description?: string
   thumbnail_url?: string
   title?: string
   updated_at?: string
+  version?: string
 }
 
 export interface ModelProductMetadataEmojis {
@@ -1148,6 +1153,7 @@ export interface ResponseAssetToken {
   id?: string
   name?: string
   native?: boolean
+  pnl?: string
   price?: number
   symbol?: string
 }
@@ -2305,6 +2311,7 @@ export interface ResponseProductBotCommand {
 
 export interface ResponseProductChangelogs {
   data?: ModelProductChangelogs[]
+  pagination?: ResponsePaginationResponse
 }
 
 export interface ResponseProfileAirdropCampaignResponse {